Add the date parsing
It's ugly but Chrome saves in the Bookmarks JSON the date as the number
of microseconds since the 1st of January 1601 (!!)
As you may imagine, this has some consequences because the
time.Duration can express at maximum ~290 years.
This is the reason of that terrible for loop in the timeParse function.